+++ CHANGELOG.md
@@ -4,11 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-
-## [0.0.1] - 2019-10-07
+## [0.0.1] - 2019-11-20
 
 ### Added
+- test: add test to cover RepositoryFinder use cases
+- ci: add ci that run test and deploy documentation
+- feat: minimal interfaces and implementations
 - docs: .gitignore to ignore unnecessary files
 - docs: CHANGELOG file to follow every changes of this project
 - docs: README file to introduce and explain this project
+
+## [Unreleased]